+++ src/GRA.Controllers/MissionControl/EventsController.cs
@@ -9,6 +9,8 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Threading.Tasks;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace GRA.Controllers.MissionControl
 {
@@ -31,9 +33,40 @@ namespace GRA.Controllers.MissionControl
             PageTitle = "Events";
         }
 
-        public async Task<IActionResult> Index(int page = 1)
+        public async Task<IActionResult> Index(string search,
+            int? systemId, int? branchId, bool? mine, int? programId, int page = 1)
         {
             BaseFilter filter = new BaseFilter(page);
+
+            if (!string.IsNullOrWhiteSpace(search))
+            {
+                filter.Search = search;
+            }
+
+            if (mine == true)
+            {
+                filter.UserIds = new List<int>() { GetId(ClaimType.UserId) };
+            }
+            else if (branchId.HasValue)
+            {
+                filter.BranchIds = new List<int>() { branchId.Value };
+            }
+            else if (systemId.HasValue)
+            {
+                filter.SystemIds = new List<int>() { systemId.Value };
+            }
+
+            if (programId.HasValue)
+            {
+                if (programId.Value > 0)
+                {
+                    filter.ProgramIds = new List<int?>() { programId.Value };
+                }
+                else
+                {
+                    filter.ProgramIds = new List<int?>() { null };
+                }
+            }
 
             var eventList = await _eventService.GetPaginatedListAsync(filter, true);
 
@@ -43,22 +76,74 @@ namespace GRA.Controllers.MissionControl
                 CurrentPage = page,
                 ItemsPerPage = filter.Take.Value
             };
-
             if (paginateModel.MaxPage > 0 && paginateModel.CurrentPage > paginateModel.MaxPage)
             {
-                return RedirectToRoute(
-                    new
-                    {
-                        page = paginateModel.LastPage ?? 1
-                    });
+                return RedirectToRoute(new { page = paginateModel.LastPage ?? 1 });
             }
+
+            var systemList = (await _siteService.GetSystemList())
+                .OrderByDescending(_ => _.Id == GetId(ClaimType.SystemId)).ThenBy(_ => _.Name);
 
             EventsListViewModel viewModel = new EventsListViewModel()
             {
                 Events = eventList.Data,
                 PaginateModel = paginateModel,
+                Search = search,
+                SystemId = systemId,
+                BranchId = branchId,
+                ProgramId = programId,
+                Mine = mine,
+                SystemList = systemList,
+                ProgramList = await _siteService.GetProgramList(),
                 CanManageLocations = UserHasPermission(Permission.ManageLocations)
             };
+
+            if (mine == true)
+            {
+                viewModel.BranchList = (await _siteService.GetBranches(GetId(ClaimType.SystemId)))
+                        .OrderByDescending(_ => _.Id == GetId(ClaimType.BranchId))
+                        .ThenBy(_ => _.Name);
+                viewModel.ActiveNav = "Mine";
+            }
+            else if (branchId.HasValue)
+            {
+                var branch = await _siteService.GetBranchByIdAsync(branchId.Value);
+                viewModel.BranchName = branch.Name;
+                viewModel.SystemName = systemList
+                    .Where(_ => _.Id == branch.SystemId).SingleOrDefault().Name;
+                viewModel.BranchList = (await _siteService.GetBranches(branch.SystemId))
+                    .OrderByDescending(_ => _.Id == GetId(ClaimType.BranchId))
+                    .ThenBy(_ => _.Name);
+                viewModel.ActiveNav = "Branch";
+            }
+            else if (systemId.HasValue)
+            {
+                viewModel.SystemName = systemList
+                    .Where(_ => _.Id == systemId.Value).SingleOrDefault().Name;
+                viewModel.BranchList = (await _siteService.GetBranches(systemId.Value))
+                    .OrderByDescending(_ => _.Id == GetId(ClaimType.BranchId))
+                    .ThenBy(_ => _.Name);
+                viewModel.ActiveNav = "System";
+            }
+            else
+            {
+                viewModel.BranchList = (await _siteService.GetBranches(GetId(ClaimType.SystemId)))
+                        .OrderByDescending(_ => _.Id == GetId(ClaimType.BranchId))
+                        .ThenBy(_ => _.Name);
+                viewModel.ActiveNav = "All";
+            }
+            if (programId.HasValue)
+            {
+                if (programId.Value > 0)
+                {
+                    viewModel.ProgramName =
+                        (await _siteService.GetProgramByIdAsync(programId.Value)).Name;
+                }
+                else
+                {
+                    viewModel.ProgramName = "Not Limited";
+                }
+            }
 
             return View(viewModel);
         }