+++ src/GRA.Controllers/ViewModel/MissionControl/Participants/ParticipantsListViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using GRA.Controllers.ViewModel.Shared;
 using System.Collections.Generic;
+using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace GRA.Controllers.ViewModel.MissionControl.Participants
 {
@@ -8,8 +9,22 @@ namespace GRA.Controllers.ViewModel.MissionControl.Participants
         public IEnumerable<GRA.Domain.Model.User> Users { get; set; }
         public PaginateViewModel PaginateModel { get; set; }
         public string Search { get; set; }
+        public string Sort { get; set; }
+        public bool IsDescending { get; set; }
+        public int? SystemId { get; set; }
+        public int? BranchId { get; set; }
+        public int? ProgramId { get; set; }
+        public string ActiveNav { get; set; }
+        public string SystemName { get; set; }
+        public string BranchName { get; set; }
+        public string ProgramName { get; set; }
         public bool CanRemoveParticipant { get; set; }
         public bool CanViewDetails { get; set; }
         public System.Array SortUsers { get; set; }
+        
+
+        public IEnumerable<GRA.Domain.Model.Branch> BranchList { get; set; }
+        public IEnumerable<GRA.Domain.Model.System> SystemList { get; set; }
+        public IEnumerable<GRA.Domain.Model.Program> ProgramList { get; set; }
     }
 }